Extract HeaderActionButton, TodoAlertDialog, and KillConfirm out of Pond.tsx.
Pond.tsx goes from 2334 to 1820 lines. The three extracted pieces are
self-contained widgets with clear boundaries:

- HeaderActionButton.tsx (105 lines): generic icon-button + hover
  tooltip.
- TodoAlertDialog.tsx (241 lines): the popup, its OnOffSwitch, the
  popover focus-trap hook, and a point-in-convex-polygon helper.
  Exposes isDialogKeyboardActive() so Pond's command-mode handler
  can bail while the dialog owns a/t.
- KillConfirm.tsx (195 lines): KillConfirmCard, KillConfirmOverlay,
  orchestrateKill, randomKillChar, and the ConfirmKill type.
