+++ lib/src/components/Pond.tsx
@@ -14,7 +14,10 @@ import { createPortal } from 'react-dom';
 import { TerminalPane } from './TerminalPane';
 import { Baseboard } from './Baseboard';
 import { tv } from 'tailwind-variants';
-import { PopupButtonRow, popupButton, renderShortcuts, Shortcut } from './design';
+import { PopupButtonRow, popupButton } from './design';
+import { HeaderActionButton } from './HeaderActionButton';
+import { TodoAlertDialog, isDialogKeyboardActive } from './TodoAlertDialog';
+import { KillConfirmOverlay, orchestrateKill, randomKillChar, type ConfirmKill } from './KillConfirm';
 import { BellIcon, BellSlashIcon, SplitHorizontalIcon, SplitVerticalIcon, ArrowsOutIcon, ArrowsInIcon, ArrowLineDownIcon, XIcon, CursorClickIcon, SelectionSlashIcon } from '@phosphor-icons/react';
 import {
   DEFAULT_MOUSE_SELECTION_STATE,
@@ -33,17 +36,13 @@ import {
   clearSessionAttention,
   clearSessionTodo,
   DEFAULT_ACTIVITY_STATE,
-  disableSessionAlert,
   dismissOrToggleAlert,
   focusSession,
   getActivity,
   getActivitySnapshot,
   markSessionAttention,
-  markSessionTodo,
   subscribeToActivity,
-  toggleSessionAlert,
   toggleSessionTodo,
-  disposeSession,
   swapTerminals,
   setPendingShellOpts,
   getDefaultShellOpts,
@@ -68,19 +67,11 @@ const mousetermTheme: DockviewTheme = {
   dndPanelOverlay: 'group',
 };
 
-let dialogKeyboardActive = false;
-
 // --- Types ---
 
 export type DooredItem = Omit<PersistedDoor, 'layoutAtMinimize'> & {
   layoutAtMinimize: SerializedDockview | null;
 };
-
-interface ConfirmKill {
-  id: string;
-  char: string;
-  shaking?: boolean;
-}
 
 export type PondMode = 'command' | 'passthrough';
 
@@ -104,108 +95,6 @@ const tabVariant = tv({
     },
   },
 });
-
-interface HeaderActionButtonProps {
-  className: string;
-  ariaLabel: string;
-  tooltip?: string;
-  tooltipDetail?: string;
-  tooltipAlign?: 'left' | 'right';
-  onMouseDownCapture?: (e: React.MouseEvent<HTMLButtonElement>) => void;
-  onMouseDown?: (e: React.MouseEvent<HTMLButtonElement>) => void;
-  onClick: (e: React.MouseEvent<HTMLButtonElement>) => void;
-  onContextMenu?: (e: React.MouseEvent<HTMLButtonElement>) => void;
-  children: React.ReactNode;
-  dataAlertButtonFor?: string;
-}
-
-function HeaderActionButton({
-  className,
-  ariaLabel,
-  tooltip,
-  tooltipDetail,
-  tooltipAlign = 'right',
-  onMouseDownCapture,
-  onMouseDown,
-  onClick,
-  onContextMenu,
-  children,
-  dataAlertButtonFor,
-}: HeaderActionButtonProps) {
-  const buttonRef = useRef<HTMLButtonElement>(null);
-  const [isVisible, setIsVisible] = useState(false);
-  const [tooltipStyle, setTooltipStyle] = useState<React.CSSProperties | null>(null);
-  const tooltipPrimary = tooltip ?? ariaLabel;
-
-  useEffect(() => {
-    if (!isVisible || !buttonRef.current) return;
-
-    const updatePosition = () => {
-      const rect = buttonRef.current?.getBoundingClientRect();
-      if (!rect) return;
-      setTooltipStyle({
-        position: 'fixed',
-        left: tooltipAlign === 'left' ? rect.left : rect.right,
-        top: rect.bottom + 8,
-        transform: tooltipAlign === 'left' ? 'translate(0, 0)' : 'translate(-100%, 0)',
-      });
-    };
-
-    updatePosition();
-    window.addEventListener('scroll', updatePosition, true);
-    window.addEventListener('resize', updatePosition);
-    return () => {
-      window.removeEventListener('scroll', updatePosition, true);
-      window.removeEventListener('resize', updatePosition);
-    };
-  }, [isVisible, tooltipAlign]);
-
-  return (
-    <>
-    <div className="relative flex shrink-0 items-center">
-      <button
-        ref={buttonRef}
-        type="button"
-        className={className}
-        data-alert-button-for={dataAlertButtonFor}
-        onMouseDownCapture={onMouseDownCapture}
-        onMouseDown={(e) => {
-          e.preventDefault();
-          e.stopPropagation();
-          onMouseDown?.(e);
-        }}
-        onClick={(e) => {
-          e.stopPropagation();
-          onClick(e);
-        }}
-        onContextMenu={onContextMenu ? (e) => {
-          e.preventDefault();
-          e.stopPropagation();
-          onContextMenu(e);
-        } : undefined}
-        aria-label={ariaLabel}
-        onMouseEnter={() => setIsVisible(true)}
-        onMouseLeave={() => setIsVisible(false)}
-      >
-        {children}
-      </button>
-    </div>
-    {isVisible && tooltipStyle && createPortal(
-      <PopupButtonRow
-        role="tooltip"
-        className="pointer-events-none z-[9999] whitespace-nowrap px-2 py-1.5"
-        style={tooltipStyle}
-      >
-        <div className="flex flex-col gap-0.5 leading-none">
-          <div>{renderShortcuts(tooltipPrimary)}</div>
-          {tooltipDetail && <div>{renderShortcuts(tooltipDetail)}</div>}
-        </div>
-      </PopupButtonRow>,
-      document.body,
-    )}
-    </>
-  );
-}
 
 // --- Alert context menu (right-click on bell) ---
 
@@ -274,19 +163,6 @@ function MouseOverrideBanner({
   );
 }
 
-function pointInConvexPolygon(x: number, y: number, vertices: Array<{ x: number; y: number }>): boolean {
-  let sign = 0;
-  for (let i = 0; i < vertices.length; i++) {
-    const a = vertices[i];
-    const b = vertices[(i + 1) % vertices.length];
-    const cross = (b.x - a.x) * (y - a.y) - (b.y - a.y) * (x - a.x);
-    if (cross === 0) continue;
-    if (sign === 0) sign = cross > 0 ? 1 : -1;
-    else if ((cross > 0 ? 1 : -1) !== sign) return false;
-  }
-  return true;
-}
-
 function clampOverlayPosition({ left, top, width, height }: {
   left: number;
   top: number;
@@ -302,210 +178,6 @@ function clampOverlayPosition({ left, top, width, height }: {
     left: Math.min(Math.max(left, margin), maxLeft),
     top: Math.min(Math.max(top, margin), maxTop),
   };
-}
-
-/**
- * Manages focus trapping, Escape-to-close, and click-outside-to-close for
- * portal-based popovers. Scopes keyboard handling to the popover's DOM subtree
- * so Tab/Escape don't leak to the rest of the app.
- */
-function usePopoverFocusTrap(
-  ref: React.RefObject<HTMLElement | null>,
-  onClose: () => void,
-  restoreFocusSelector?: string,
-) {
-  useEffect(() => {
-    const el = ref.current;
-    if (!el) return;
-
-    const handleMouseDown = (e: MouseEvent) => {
-      if (!el.contains(e.target as Node)) onClose();
-    };
-
-    const handleKeyDown = (e: KeyboardEvent) => {
-      // Only handle keys when focus is inside the popover
-      if (!el.contains(document.activeElement)) return;
-
-      if (e.key === 'Escape') {
-        e.preventDefault();
-        e.stopPropagation();
-        onClose();
-        return;
-      }
-      if (e.key !== 'Tab') return;
-
-      const focusables = Array.from(
-        el.querySelectorAll<HTMLElement>('button:not([disabled]), [tabindex]:not([tabindex="-1"])'),
-      );
-      if (focusables.length === 0) return;
-
-      const currentIndex = focusables.findIndex((f) => f === document.activeElement);
-      const nextIndex = currentIndex === -1
-        ? 0
-        : (currentIndex + (e.shiftKey ? -1 : 1) + focusables.length) % focusables.length;
-
-      e.preventDefault();
-      focusables[nextIndex]?.focus();
-    };
-
-    window.addEventListener('mousedown', handleMouseDown);
-    window.addEventListener('keydown', handleKeyDown, true);
-    return () => {
-      window.removeEventListener('mousedown', handleMouseDown);
-      window.removeEventListener('keydown', handleKeyDown, true);
-      if (restoreFocusSelector) {
-        document.querySelector<HTMLElement>(restoreFocusSelector)?.focus();
-      }
-    };
-  }, [ref, onClose, restoreFocusSelector]);
-}
-
-function TodoAlertDialog({
-  position,
-  sessionId,
-  onClose,
-}: {
-  position: { x: number; y: number; triggerRect: DOMRect };
-  sessionId: string;
-  onClose: () => void;
-}) {
-  const activityStates = useSyncExternalStore(subscribeToActivity, getActivitySnapshot);
-  const activity = activityStates.get(sessionId) ?? DEFAULT_ACTIVITY_STATE;
-  const alertEnabled = activity.status !== 'ALERT_DISABLED';
-  const dialogRef = useRef<HTMLDivElement>(null);
-
-  usePopoverFocusTrap(dialogRef, onClose, `[data-alert-button-for="${sessionId}"]`);
-
-  useEffect(() => {
-    dialogRef.current?.querySelector<HTMLElement>('button')?.focus();
-  }, []);
-
-  // Keyboard shortcuts within dialog
-  useEffect(() => {
-    const el = dialogRef.current;
-    if (!el) return;
-    dialogKeyboardActive = true;
-    const handler = (e: KeyboardEvent) => {
-      if (!el.contains(document.activeElement)) return;
-      if (e.key === 'a') {
-        e.preventDefault();
-        e.stopImmediatePropagation();
-        dismissOrToggleAlert(sessionId, getActivity(sessionId).status);
-      }
-      if (e.key === 't') {
-        e.preventDefault();
-        e.stopImmediatePropagation();
-        toggleSessionTodo(sessionId);
-      }
-    };
-    window.addEventListener('keydown', handler, true);
-    return () => {
-      dialogKeyboardActive = false;
-      window.removeEventListener('keydown', handler, true);
-    };
-  }, [sessionId]);
-
-  // Hot area: close when mouse leaves (dialog ∪ funnel from trigger button to dialog top).
-  useEffect(() => {
-    const dialog = dialogRef.current;
-    if (!dialog) return;
-    const trigger = position.triggerRect;
-    const handler = (e: MouseEvent) => {
-      const dialogRect = dialog.getBoundingClientRect();
-      const { clientX: x, clientY: y } = e;
-      if (x >= dialogRect.left && x <= dialogRect.right && y >= dialogRect.top && y <= dialogRect.bottom) return;
-      const funnel = [
-        { x: trigger.left, y: trigger.top },
-        { x: trigger.right, y: trigger.top },
-        { x: dialogRect.right, y: dialogRect.top },
-        { x: dialogRect.left, y: dialogRect.top },
-      ];
-      if (pointInConvexPolygon(x, y, funnel)) return;
-      onClose();
-    };
-    window.addEventListener('mousemove', handler);
-    return () => window.removeEventListener('mousemove', handler);
-  }, [position.triggerRect, onClose]);
-
-  return createPortal(
-    <div
-      ref={dialogRef}
-      className="fixed z-[9999] w-fit rounded-lg border border-border bg-surface-raised p-3 shadow-lg"
-      style={{ left: position.x, top: position.y }}
-      role="dialog"
-      aria-modal="true"
-      aria-label="TODO and alert settings"
-    >
-      <button
-        type="button"
-        aria-label="Close"
-        className="absolute right-2 top-2 rounded p-0.5 text-muted hover:bg-foreground/10 hover:text-foreground"
-        onClick={onClose}
-      >
-        <XIcon size={12} weight="bold" />
-      </button>
-
-      <div className="mb-3 grid w-fit grid-cols-[auto_auto_auto] items-center gap-x-2 gap-y-2">
-        {/* TODO row */}
-        <Shortcut>t</Shortcut>
-        <span className="text-xs font-medium text-foreground">TODO</span>
-        <OnOffSwitch
-          on={activity.todo}
-          onEnable={() => markSessionTodo(sessionId)}
-          onDisable={() => clearSessionTodo(sessionId)}
-          label="TODO"
-        />
-
-        {/* Alert row */}
-        <Shortcut>a</Shortcut>
-        <span className="text-xs font-medium text-foreground">alert</span>
-        <OnOffSwitch
-          on={alertEnabled}
-          onEnable={() => toggleSessionAlert(sessionId)}
-          onDisable={() => disableSessionAlert(sessionId)}
-          label="alert"
-        />
-      </div>
-
-      <div className="border-t border-border pt-2 text-xs leading-relaxed text-muted">
-        When a tab with a ringing alert is selected,<br />
-        the alert is cleared and the tab gets a TODO.<br />
-        Pressing [Enter] into the tab will clear the TODO.
-      </div>
-    </div>,
-    document.body,
-  );
-}
-
-function OnOffSwitch({
-  on,
-  onEnable,
-  onDisable,
-  label,
-}: {
-  on: boolean;
-  onEnable: () => void;
-  onDisable: () => void;
-  label: string;
-}) {
-  return (
-    <button
-      type="button"
-      role="switch"
-      aria-checked={on}
-      aria-label={`${label} ${on ? 'on' : 'off'}`}
-      onClick={() => (on ? onDisable() : onEnable())}
-      className="relative inline-flex h-5 w-14 items-center rounded-full border border-border bg-surface text-xs font-medium"
-    >
-      <span
-        aria-hidden
-        className="absolute inset-y-0.5 w-[calc(50%-2px)] rounded-full bg-accent/25 transition-transform"
-        style={{ transform: on ? 'translateX(2px)' : 'translateX(calc(100% + 2px))' }}
-      />
-      <span className={['z-10 flex-1 text-center', on ? 'text-accent' : 'text-muted'].join(' ')}>on</span>
-      <span className={['z-10 flex-1 text-center', on ? 'text-muted' : 'text-accent'].join(' ')}>off</span>
-    </button>
-  );
 }
 
 // --- Contexts ---
@@ -585,12 +257,6 @@ function idsMatch(a: string[], b: string[]): boolean {
     console.assert(isSorted(a) && isSorted(b), 'idsMatch: inputs must be sorted');
   }
   return a.length === b.length && a.every((id, i) => id === b[i]);
-}
-
-/** Random A-Z excluding X (prevents accidental double-tap on kill shortcut) */
-const KILL_CONFIRM_CHARS = 'ABCDEFGHIJKLMNOPQRSTUVWYZ'; // no X
-function randomKillChar(): string {
-  return KILL_CONFIRM_CHARS[Math.floor(Math.random() * KILL_CONFIRM_CHARS.length)];
 }
 
 // --- Panel content component ---
@@ -1097,186 +763,6 @@ function SelectionOverlay({ apiRef, selectedId, selectedType, mode, overlayElRef
   );
 }
 
-// --- Kill confirmation overlay ---
-
-export function KillConfirmCard({ char, onCancel, shaking }: { char: string; onCancel?: () => void; shaking?: boolean }) {
-  return (
-    <div className={`bg-surface-raised border border-error/30 px-6 py-4 rounded-lg text-center shadow-lg${shaking ? ' motion-safe:animate-shake-x' : ''}`}>
-      <h2 className="text-base font-bold mb-3 text-foreground">Kill Session?</h2>
-      <div className="bg-black py-2 px-6 rounded border border-border inline-block mb-2">
-        <span className="text-xl font-bold text-error">{char}</span>
-      </div>
-      <div className="text-xs text-muted uppercase tracking-widest leading-relaxed">
-        <div>[{char}] to confirm</div>
-        <button type="button" onClick={onCancel} className="uppercase hover:text-foreground transition-colors cursor-pointer">[ESC] to cancel</button>
-      </div>
-    </div>
-  );
-}
-
-function KillConfirmOverlay({ confirmKill, panelElements, onCancel }: {
-  confirmKill: ConfirmKill;
-  panelElements: Map<string, HTMLElement>;
-  onCancel: () => void;
-}) {
-  const [rect, setRect] = useState<{ top: number; left: number; width: number; height: number } | null>(null);
-
-  // useLayoutEffect (not useEffect) so the initial measurement + re-render happens
-  // before the browser paints. Otherwise the centered-in-viewport fallback below
-  // flashes for one frame before the overlay snaps to the panel.
-  useLayoutEffect(() => {
-    const panelEl = resolvePanelElement(panelElements.get(confirmKill.id));
-    if (!panelEl) { setRect(null); return; }
-
-    const update = () => {
-      const r = panelEl.getBoundingClientRect();
-      setRect({ top: r.top, left: r.left, width: r.width, height: r.height });
-    };
-
-    update();
-    const ro = new ResizeObserver(update);
-    ro.observe(panelEl);
-    window.addEventListener('resize', update);
-    return () => { ro.disconnect(); window.removeEventListener('resize', update); };
-  }, [confirmKill.id, panelElements]);
-
-  if (rect) {
-    return (
-      <div
-        style={{ position: 'fixed', top: rect.top, left: rect.left, width: rect.width, height: rect.height, zIndex: 100 }}
-        className="flex items-center justify-center bg-surface/50 rounded"
-      >
-        <KillConfirmCard char={confirmKill.char} onCancel={onCancel} shaking={confirmKill.shaking} />
-      </div>
-    );
-  }
-
-  // Fallback: centered in viewport
-  return (
-    <div className="fixed inset-0 bg-surface/50 z-[100] flex items-center justify-center">
-      <KillConfirmCard char={confirmKill.char} onCancel={onCancel} shaking={confirmKill.shaking} />
-    </div>
-  );
-}
-
-
-// --- Kill animation ---
-//
-// Orchestrates the visual reclaim when a pane is killed:
-//   1. Fade the real killed pane's group element in place (its actual content
-//      dissolves — a solid-color ghost over a same-colored background would be
-//      invisible).
-//   2. After the fade completes, capture pre-rects of surviving panes, remove
-//      the panel (dockview snaps the layout), and FLIP each grower via
-//      clip-path so its newly claimed territory is hidden at start and swept
-//      in by the transition. clip-path (not transform) keeps
-//      getBoundingClientRect accurate so the SelectionOverlay doesn't lag.
-//
-// killInProgressRef is set across api.removePanel so the onDidRemovePanel
-// auto-spawn handler knows we already waited for our own fade and can skip
-// its own 440ms delay (avoids stacking 440ms + 440ms on last-pane kill).
-function orchestrateKill(
-  api: DockviewApi,
-  killedId: string,
-  selectPanel: (id: string) => void,
-  setSelectedId: (id: string | null) => void,
-  killInProgressRef: { current: boolean },
-  overlayElRef: { current: HTMLElement | null },
-): void {
-  const panel = api.getPanel(killedId);
-  if (!panel) return;
-
-  const bareRemove = () => {
-    killInProgressRef.current = true;
-    disposeSession(killedId);
-    api.removePanel(panel);
-    killInProgressRef.current = false;
-    if (api.panels.length > 0) selectPanel(api.panels[0].id);
-    else setSelectedId(null);
-  };
-
-  const reduceMotion = typeof window !== 'undefined'
-    && window.matchMedia?.('(prefers-reduced-motion: reduce)').matches;
-  const killedGroupEl = panel.api.group?.element;
-  if (reduceMotion || !killedGroupEl) {
-    bareRemove();
-    return;
-  }
-
-  // Fade the killed pane in place. Block input on it during the fade.
-  // For a last-pane kill (auto-spawn will create a replacement), also shrink
-  // the pane toward the bottom-right so the disappearance is visible — a plain
-  // fade offers no visual cue since the pane's space is reclaimed by a new one
-  // appearing in exactly the same rect from the opposite corner. The focus
-  // ring (SelectionOverlay element) gets a matching shrink animation so it
-  // scales with the pane rather than sitting over empty space.
-  const isLastPane = api.panels.length === 1;
-  const fadeClass = isLastPane ? 'pane-fading-and-shrinking-to-br' : 'pane-fading-out';
-  const fadeAnimationName = isLastPane ? 'pane-fade-and-shrink-to-br' : 'pane-fade-out';
-  killedGroupEl.style.pointerEvents = 'none';
-  killedGroupEl.classList.add(fadeClass);
-  const overlayEl = isLastPane ? overlayElRef.current : null;
-  if (overlayEl) overlayEl.classList.add('ring-shrinking-to-br');
-
-  let finalized = false;
-  const finalize = () => {
-    if (finalized) return;
-    finalized = true;
-
-    // Snapshot pre-rects just before removal.
-    interface Pre { el: HTMLElement; rect: DOMRect; }
-    const preRects = new Map<string, Pre>();
-    for (const p of api.panels) {
-      if (p.id === killedId) continue;
-      const el = p.api.group?.element;
-      if (el) preRects.set(p.id, { el, rect: el.getBoundingClientRect() });
-    }
-
-    bareRemove();
-
-    // FLIP each grower.
-    for (const p of api.panels) {
-      const pre = preRects.get(p.id);
-      if (!pre) continue;
-      const postRect = pre.el.getBoundingClientRect();
-      const dw = postRect.width - pre.rect.width;
-      const dh = postRect.height - pre.rect.height;
-      if (Math.abs(dw) < 0.5 && Math.abs(dh) < 0.5) continue;
-
-      // Clear any in-progress spawn animation before applying FLIP.
-      pre.el.classList.remove('pane-spawning-from-left', 'pane-spawning-from-top', 'pane-spawning-from-top-left');
-
-      const clipTop    = Math.max(0, (pre.rect.top - postRect.top)       / postRect.height * 100);
-      const clipBottom = Math.max(0, (postRect.bottom - pre.rect.bottom) / postRect.height * 100);
-      const clipLeft   = Math.max(0, (pre.rect.left - postRect.left)     / postRect.width  * 100);
-      const clipRight  = Math.max(0, (postRect.right - pre.rect.right)   / postRect.width  * 100);
-
-      pre.el.style.transition = 'none';
-      pre.el.style.clipPath = `inset(${clipTop}% ${clipRight}% ${clipBottom}% ${clipLeft}%)`;
-      void pre.el.offsetHeight;
-      pre.el.style.transition = 'clip-path 440ms cubic-bezier(0.22, 1, 0.36, 1)';
-      pre.el.style.clipPath = 'inset(0)';
-      const cleanup = () => {
-        pre.el.style.transition = '';
-        pre.el.style.clipPath = '';
-      };
-      pre.el.addEventListener('transitionend', cleanup, { once: true });
-      setTimeout(cleanup, 1000);
-    }
-
-    // Peel the ring-shrink class so the next selection's overlay renders at
-    // full scale. The element may have been reused by React for the next
-    // selected pane's overlay by the time the animation finishes.
-    if (overlayEl) overlayEl.classList.remove('ring-shrinking-to-br');
-  };
-
-  killedGroupEl.addEventListener('animationend', (ev) => {
-    if ((ev as AnimationEvent).animationName !== fadeAnimationName) return;
-    finalize();
-  });
-  // Safety: if animationend never fires, still finalize.
-  setTimeout(finalize, 1000);
-}
 
 
 // --- Main component ---
@@ -1994,7 +1480,7 @@ export function Pond({
       }
 
       if (e.key === 't' && sid && selectedTypeRef.current === 'pane') {
-        if (dialogKeyboardActive) return;
+        if (isDialogKeyboardActive()) return;
         e.preventDefault();
         e.stopPropagation();
         toggleSessionTodo(sid);
@@ -2002,7 +1488,7 @@ export function Pond({
       }
 
       if (e.key === 'a' && sid && selectedTypeRef.current === 'pane') {
-        if (dialogKeyboardActive) return;
+        if (isDialogKeyboardActive()) return;
         e.preventDefault();
         e.stopPropagation();
         dismissOrToggleAlert(sid, getActivity(sid).status);
